+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2017-06-26
+### Added
+* WINDOWS_INSTALLATION.md
+
+### Changed
+* Dictionary {'θ':'th' => 'θ':'8', 'ξ':'ks' => 'ξ':'3'}
+* README.md
+
 ## [1.0.1] - 2017-06-24
 ### Added
+* Translate any language into Greeklish (type + enter translation ONLY for now)
+* Translate files contain Greek into Greeklish (write new files, optional)
 * LICENSE, CHANGELOG, requirements files
 
 ### Changed